Refactor: percent Int to Float 및 그래프 데이터 수정

+++ core/data/src/main/java/kr/co/data/model/data/AccountBookListData.kt
@@ -25,7 +25,7 @@ data class AccountBookListData(
     )
 
     data class PercentCategory(
-        val percent: Int,
+        val percent: Float,
         val category: String
     )
 }
